Implement authentication.
Wraps the app in a wrapper that listens for auth changes.
Wrapper listens for auth changes and displays either the authentication
widget or the charger-category widget if the user is logged out/logged
in respectively.

Swipe left on the charger-category page to open a drawer containing the sign out button.