Fix issue with anonymous sign in and geolocator.
Anonymous sign in does not initialize displayName field in the database,
which results in a runtime error when the database Service attempts to
retrieve the displayName field.

Geolocator.getLastKnownPosition returns null on emulator devices if
device's lastKnownPosition is not initialized. Added null check as a
precaution for emulator and real devices.